+++ ibis/expr/operations/arrays.py
@@ -4,11 +4,11 @@ from ibis.common import exceptions as com
 from ibis.common.validators import immutable_property
 from ibis.expr import datatypes as dt
 from ibis.expr import rules as rlz
-from ibis.expr.operations.core import UnaryOp, ValueOp
+from ibis.expr.operations.core import UnaryOp, Value
 
 
 @public
-class ArrayColumn(ValueOp):
+class ArrayColumn(Value):
     cols = rlz.value_list_of(rlz.column(rlz.any), min_length=1)
 
     output_shape = rlz.Shape.COLUMNAR
@@ -36,7 +36,7 @@ class ArrayLength(UnaryOp):
 
 
 @public
-class ArraySlice(ValueOp):
+class ArraySlice(Value):
     arg = rlz.array
     start = rlz.integer
     stop = rlz.optional(rlz.integer)
@@ -46,7 +46,7 @@ class ArraySlice(ValueOp):
 
 
 @public
-class ArrayIndex(ValueOp):
+class ArrayIndex(Value):
     arg = rlz.array
     index = rlz.integer
 
@@ -58,7 +58,7 @@ class ArrayIndex(ValueOp):
 
 
 @public
-class ArrayConcat(ValueOp):
+class ArrayConcat(Value):
     left = rlz.array
     right = rlz.array
 
@@ -78,7 +78,7 @@ class ArrayConcat(ValueOp):
 
 
 @public
-class ArrayRepeat(ValueOp):
+class ArrayRepeat(Value):
     arg = rlz.array
     times = rlz.integer
 
@@ -87,7 +87,7 @@ class ArrayRepeat(ValueOp):
 
 
 @public
-class Unnest(ValueOp):
+class Unnest(Value):
     arg = rlz.array
 
     @immutable_property